+++ requirements-dev.txt
@@ -25,6 +25,7 @@ pandas==1.3.5; python_version < '3.8'
 pandas==2.0.2; python_version >= '3.8'
 xxhash>=3.0.0
 Pillow==9.5.0
+opencv-python==4.7.0.72
 
 # Ray
 ray[data, default]==2.4.0
